+++ src/control_plane/shared_infrastructure/postgres/mod.rs
@@ -2,8 +2,10 @@ mod postgres_logical_resource_plan;
 mod postgres_plan_error;
 mod postgres_shared_instance_plan;
 mod postgres_shared_instance_plan_options;
+mod provision_postgres_logical_resource;
 
 pub(crate) use postgres_logical_resource_plan::PostgresLogicalResourcePlan;
 pub(crate) use postgres_plan_error::PostgresPlanError;
 pub(crate) use postgres_shared_instance_plan::PostgresSharedInstancePlan;
 pub(crate) use postgres_shared_instance_plan_options::PostgresSharedInstancePlanOptions;
+pub(crate) use provision_postgres_logical_resource::provision_postgres_logical_resource;